Create GPA Calculator Script
I focused entirely on readability, using comments to break up the sections, keeping everything grouped together with clearly defined spaces, and keeping the length of code to the same size by truncating some words like computerScienceGradeValue to compSciGradeValue.

It's bad practice to abbreviate or truncate code, but in this case - by the context of the rest of the code, and the clear readability of everything - It's easy to immediately deduce what the truncated word is implying. Whilst also adding to the neatness of the grouped section.